enhancement: migrated the common migrate code out of make and sync files
