+++ lib/migration.js
@@ -1,10 +1,9 @@
-import { writeFileSync } from 'fs';
-import { join } from 'path';
-
+import fs, { writeFileSync } from 'fs';
+import path, { join } from 'path';
 import prettier from 'prettier';
-
 import { getYYYYMMDDHHMMSS, template } from './functions.js';
-import { readAsset } from './helpers.js';
+import { readAsset, sortActions } from './helpers.js';
+import { parseDifference, reverseModels } from './models.js';
 
 const { log } = console;
 
@@ -135,6 +134,67 @@ export const getPartialMigration = (actions) => {
   });
 
   return { commands, consoles };
+};
+
+export const migrate = async (options) => {
+  const { modelsDir, migrationsDir, stateDir, indexDir } = options;
+
+  if (!fs.existsSync(modelsDir)) {
+    throw "Can't find models directory. Use `sequelize init` to create it";
+  }
+
+  if (!fs.existsSync(migrationsDir)) {
+    throw "Can't find migrations directory. Use `sequelize init` to create it";
+  }
+
+  if (!fs.existsSync(stateDir)) {
+    console.log("Can't find State directory. Creating it...");
+    fs.mkdirSync(stateDir, { recursive: true });
+  }
+
+  // current state
+  const currentState = {
+    tables: {},
+    path: path.join(stateDir, '_current.json'),
+    backupPath: path.join(stateDir, '_current_bak.json'),
+  };
+
+  currentState.exists = fs.existsSync(currentState.path);
+
+  // load last state
+  let previousState = {
+    revision: 0,
+    tables: {},
+  };
+
+  if (currentState.exists) {
+    try {
+      previousState = JSON.parse(fs.readFileSync(currentState.path));
+    } catch (e) {
+      console.log('_current.json syntax not valid');
+    }
+  } else {
+    console.log('_current.json not found. first time running this tool');
+  }
+
+  const { sequelize } = (await import(`file:////${indexDir}`)).default;
+  const { models } = sequelize;
+
+  currentState.tables = reverseModels(sequelize, models);
+
+  const upActions = parseDifference(previousState.tables, currentState.tables);
+  const downActions = parseDifference(currentState.tables, previousState.tables);
+
+  // sort actions
+  sortActions(upActions);
+  sortActions(downActions);
+
+  const migration = getMigration(upActions, downActions);
+  return {
+    previousState,
+    currentState,
+    migration,
+  };
 };
 
 export const getMigration = (upActions, downActions) => {
